testing params

+++ test.py
@@ -355,7 +355,7 @@ def export_saved_model(version, path, sess=None):
     feature_configs = {
         # 'x': tf.FixedLenFeature(shape=[], dtype=tf.string),
         # 'y': tf.FixedLenFeature(shape=[], dtype=tf.string)
-        'x': tf.FixedLenFeature(shape=[], dtype=tf.int32),
+        'x': tf.FixedLenFeature(shape=[], dtype=tf.int64),
         'y': tf.FixedLenFeature(shape=[], dtype=tf.float32)
     }
     serialized_example = tf.placeholder(tf.string, name="tf_example")
